cloud/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.12
boto3==1.35.49
anthropic==0.40.0
pyjwt==2.9.0
sentence-transformers==3.2.1
qdrant-client==1.12.1
pymupdf==1.24.12
python-docx==1.1.2
tiktoken==0.8.0
numpy==1.26.4
websockets==12.0
paho-mqtt==2.1.0
cryptography==42.0.0
jinja2==3.1.4
beautifulsoup4==4.12.3
duckduckgo-search==6.3.7
pytest==8.3.3
ruff==0.8.4